requirements.txt
accelerate>=0.31.0
decord>=0.6.0
ftfy
gradio>=4.0.0
imageio
imageio-ffmpeg
Jinja2
numpy
opencv-python
open-clip-torch
peft>=0.11.1
Pillow
sentencepiece
tensorboard
torchvision
transformers>=4.41.2
